Improve readability and added “Hide Ask Button”.
Previously, I added hide Hype button, next I have added the hide Ask button, but as usual it may not work because of the lack of an identifier, also updated the code’s readability to note which button has an missing or unidentified identifier. It is still possible to hide them if the identifiers to the buttons exist.